@@ -31,18 +31,18 @@ unsafe partial struct Blake2sHashState
 		var r8  = Unsafe.As<byte, Vector128<byte>>(ref Unsafe.Add(ref rrm, Vector128<byte>.Count));
 
 		var row1 = Sse2.LoadVector128(sh);
-		var row2 = Sse2.LoadVector128(sh + 4);
+		var row2 = Sse2.LoadVector128(sh + Vector128<uint>.Count);
 
 		ref byte riv = ref MemoryMarshal.GetReference(ivle);
 		var row3 = Unsafe.As<byte, Vector128<uint>>(ref riv);
 		var row4 = Unsafe.As<byte, Vector128<uint>>(ref Unsafe.Add(ref riv, 16));
 
-		row4 = Sse2.Xor(row4, Sse2.LoadVector128(sh + 8)); // t[] and f[]
+		row4 = Sse2.Xor(row4, Sse2.LoadVector128(sh + Vector128<uint>.Count * 2)); // t[] and f[]
 
 		var m0 = Sse2.LoadVector128(m);
-		var m1 = Sse2.LoadVector128(m + 4);
-		var m2 = Sse2.LoadVector128(m + 8);
-		var m3 = Sse2.LoadVector128(m + 12);
+		var m1 = Sse2.LoadVector128(m + Vector128<uint>.Count);
+		var m2 = Sse2.LoadVector128(m + Vector128<uint>.Count * 2);
+		var m3 = Sse2.LoadVector128(m + Vector128<uint>.Count * 3);
 
 		//ROUND 1
 		var b0 = Sse.Shuffle(m0.AsSingle(), m1.AsSingle(), 0b_10_00_10_00).AsUInt32();
@@ -695,9 +695,9 @@ unsafe partial struct Blake2sHashState
 		row1 = Sse2.Xor(row1, row3);
 		row2 = Sse2.Xor(row2, row4);
 		row1 = Sse2.Xor(row1, Sse2.LoadVector128(sh));
-		row2 = Sse2.Xor(row2, Sse2.LoadVector128(sh + 4));
+		row2 = Sse2.Xor(row2, Sse2.LoadVector128(sh + Vector128<uint>.Count));
 		Sse2.Store(sh, row1);
-		Sse2.Store(sh + 4, row2);
+		Sse2.Store(sh + Vector128<uint>.Count, row2);
 	}
 }
 #endif
